Started writing serialization for primitive types
